+++ NEWS.md
@@ -1,9 +1,15 @@
 
-# gge 1.0 (unpublished)
+# gge 1.1 (Sep 2016)
+
+* Added zoom.gen, zoom.env arguments to `biplot()`for M.Zoric.
+* Moved tests to tests/all.R.
+
+# gge 1.0 (Dec 2015)
 
 * Packaged forked and split off from agridat package.
+* Added origin/hull arguments to `biplot()`.
 
-# gge 0.0 ()
+# gge 0.0 (Jan 2005)
 
-* Function `gge()` to fit and plot GGE biplots.
+* Created function `gge()` to fit and plot GGE biplots.